% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFN-class.R
\name{checking_xy}
\alias{checking_xy}
\title{Checking data}
\usage{
checking_xy(object, ...)
}
\arguments{
\item{object}{SLFN object to compare the matrices X and Y}

\item{X}{a input matrix of dimensions [Nxd]}

\item{Y}{a output matrix of dimensions [Nxc]}
}
\value{
boolean value that is TRUE if everything is correct
}
\description{
\code{checking_xy} checks that inputs, outputs and dataset dimensions are correct.
Only checks the data if the variables is not NULL.
}

