% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\name{checkingXY}
\alias{checkingXY}
\title{Checking data}
\usage{
checkingXY(object, ...)
}
\arguments{
\item{object}{SLFN object to compare the matrices X and Y}

\item{X}{a input matrix of dimensions [Nxd]}

\item{Y}{a output matrix of dimensions [Nxc]}
}
\value{
boolean value that is TRUE if everything is correct
}
\description{
\code{checkingXY} checks that inputs, outputs and dataset dimensions are correct.
Only checks the data if the variables is not NULL.
}

