% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{chekingDataModel,SLFN-method}
\alias{chekingDataModel,SLFN-method}
\title{Check that the input and output data and dimensions are correct
Only checks the data if the variables is not NULL}
\usage{
\S4method{chekingDataModel}{SLFN}(object, X, Y, ...)
}
\arguments{
\item{object}{SLFN object to compare the matrices X and Y}

\item{X}{a input matrix of dimensions [Nxd]}

\item{Y}{a output matrix of dimensions [Nxc]}
}
\value{
X a input matrix of dimensions [Nxd]

Y a output matrix of dimensions [Nxc]
}
\description{
Check that the input and output data and dimensions are correct
Only checks the data if the variables is not NULL
}

