% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{computeError}
\alias{computeError}
\title{Compute model's errors with different number of neurons (nn)}
\usage{
computeError(object, ...)
}
\arguments{
\item{object}{An instance to the SLFN class.}

\item{nSelected}{The number of hidden neurons to be used.}

\item{H}{The transformed matrix H of dimensions [NxL].}

\item{Y}{The output matrix of dimensions [Nxc].}

\item{Hv}{The transformed matrix H for the validation dataset
of dimensions [NxL].}

\item{Yv}{The output matrix for the validation dataset of dimensions [Nxc].}

\item{index}{The vector containing the selection of the data.}
}
\value{
The error of the ELM model
}
\description{
\code{computeError} calculate the error of a ELM model.
}

