% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createFolds}
\alias{createFolds}
\title{createFolds from caret package
CREATE GENERIC AND TRANSFORM TO MATRICES???}
\usage{
createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\description{
createFolds from caret package
CREATE GENERIC AND TRANSFORM TO MATRICES???
}

