% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{distMatVect}
\alias{distMatVect}
\title{Distance matrix - vector}
\usage{
distMatVect(X, ref, type = "euclidean")
}
\arguments{
\item{X}{A matrix with \emph{N} samples/points from \emph{d}-space}

\item{ref}{A vector, a sample/point from \emph{d}-space}

\item{type}{A method used to compute the distance
\itemize{
\item \emph{euclidean}
\item \emph{others???}
}}
}
\value{
A vector of length \emph{N} with all distances
}
\description{
\code{distMatVect} compute the distance between each sample of matrix
(row-wise) and a reference vector.
}

