% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_neurons.R, R/elm-class.R, R/errors.R, R/predict.R, R/ranking.R, R/train.R, R/train_pruning.R
\docType{methods}
\name{add_neurons,elm-method}
\alias{add_neurons,elm-method}
\alias{class_postprocess,elm-method}
\alias{elm}
\alias{elm-class}
\alias{get_error,elm-method}
\alias{initialize,elm-method}
\alias{mse,elm-method}
\alias{project,elm-method}
\alias{prune,elm-method}
\alias{rank_neurons,elm-method}
\alias{show,elm-method}
\alias{solve_system,elm-method}
\alias{train,elm-method}
\alias{train_pruning,elm-method}
\title{Class \code{elm}}
\usage{
\S4method{add_neurons}{elm}(object, act_fun, nn, w_in = NULL, b = NULL)

\S4method{initialize}{elm}(.Object = object, inputs = 0, outputs = 0)

\S4method{show}{elm}(object)

\S4method{get_error}{elm}(object, n_sel, h, y, h_val = NULL, y_val = NULL,
  cv_rows = NULL)

\S4method{mse}{elm}(object, y, yp, x)

\S4method{class_postprocess}{elm}(object, yp, class_output, ml_threshold)

\S4method{rank_neurons}{elm}(object, nn_max, h = NULL, y = NULL)

\S4method{train}{elm}(object, x, y, x_val = NULL, y_val = NULL,
  type = "reg", tune = "none", ranking = "random", validation = "none",
  folds = 10, class_weights = NULL, ...)

\S4method{project}{elm}(object, x, rbf_dist = "euclidean")

\S4method{solve_system}{elm}(object, h, y, solve = TRUE)

\S4method{train_pruning}{elm}(object, h, y, h_val = NULL, y_val = NULL,
  cv_rows = NULL)

\S4method{prune}{elm}(object, n_sel)
}
\description{
A S4 class to represent an Extreme Learning Machine (ELM) model
}
\section{Methods (by generic)}{
\itemize{
\item \code{add_neurons}: add neurons of the same type of activation function to the hidden layer

\item \code{initialize}: initalize an object of class \code{elm}

\item \code{show}: display an object of class \code{elm}

\item \code{get_error}: implement a validation procedure

\item \code{mse}: MSE error

\item \code{class_postprocess}: Description of class_postprocess for ELM (origin predic.R)

\item \code{rank_neurons}: rank neurons of a ELM

\item \code{train}: train the elm

\item \code{project}: project form input-space to neuron-space. Compute H

\item \code{solve_system}: solve linear system H x Wout = Y

\item \code{train_pruning}: Optimization procedure for obtaining the optimial number of neurons for pruning.

\item \code{prune}: Prune the hidden layer of a elm
}}
\section{Slots}{

\describe{
\item{\code{inputs}}{The number of input features.}

\item{\code{outputs}}{The number of outputs.}

\item{\code{h_neurons}}{An object of classs hiddenlayer}

\item{\code{w_out}}{The weight output vector that includes the computed weights between
the hidden and the output layer.}

\item{\code{type}}{The type of model implemented:
\itemize{
\item "reg": regression problem.
\item "class_mc": multi-class: the sample belongs to 1 class out of n.
\item "class_ml": multi-label: the sample can belong to m classes out of n (m<n).
\item "class_w":  weigted classification
}}

\item{\code{tune}}{Parameter to define the model structure selection method implemented to tune
the model hyper-parameters
#' \itemize{
\item "none": no model selection
\item "pruning": pruning of neurons of the hidden layer: P-ELM, if "ridge = 0 &
ranking = "random", OP-ELM, if "ridge = 0 & ranking = lars", TROP-ELM, if ("ridge != 0 &
ranking = lars)
}}

\item{\code{ranking}}{A character to select the type of ranking implemented when prunning option
is selected.
\itemize{
\item "random" - random ranking
\item "lars" - ranking based on lars - L1 penalty
}}

\item{\code{results}}{The error used to evaluate model performance.
mse c(mse_train, mse_val)}

\item{\code{ridge}}{The regularization parameter used to include the L2 penalty the#'}

\item{\code{validation}}{The validation procedure used for developing the model.
#' \itemize{
\item "none" - no validation process  <<<<<<ANDRES<<<<<<
\item "v" - validation. Xv and Yv are required
\item "cv" - cross validation. The number of folds is required
\item "loo" - leave one out based on the PRESS statistic
}}

\item{\code{folds}}{The number number of folds for the cross-validation procedure.}

\item{\code{class_weights}}{numeric vector of length = number_of_classes with the weigths for weighted type}

\item{\code{batch}}{The size of the bacth in an adaptative ELM.}

\item{\code{time_exec}}{The time of calculation for training the model.}

\item{\code{bigdata}}{An logical parameter to select the kind of acceleration used in
case of solving big data problems.}
}}

