\name{elm-package}
\alias{elm-package}
\docType{package}
\title{
Exact linear models
}
\description{
Implementation of exact test of linear regressions.
}
\details{
\tabular{ll}{
Package: \tab elm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-11-17\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Karl Schlag, Olivier Gossner, Gareth Liu-Evans and Oliver Reiter
}
\references{
Olivier Gossner, Karl H. Schlag, "Finite-sample exact tests for linear regressions with bounded dependent variables", Journal of Econometrics, Volume 177, Issue 1, November 2013, Pages 75-84, ISSN 0304-4076, \url{http://dx.doi.org/10.1016/j.jeconom.2013.06.003}.
}
\seealso{
  \url{http://homepage.univie.ac.at/karl.schlag/research/statistics/}
}

