% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_clip.R
\name{elm_clip}
\alias{elm_clip}
\title{Copy the file tree to the clipboard}
\usage{
elm_clip(path = ".", dir_only = FALSE, ...)
}
\arguments{
\item{path}{the path to search}

\item{dir_only}{logical indicating whether to return only directories or both files and directories}

\item{...}{additional arguments passed to `elm_clip`}
}
\value{
nothing, copies the file tree of files to the clipboard
}
\description{
Copy the file tree to the clipboard
}
