% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_clip.R
\name{elm_clip_files}
\alias{elm_clip_files}
\title{Copy the tree structure of files and directories to the clipboard}
\usage{
elm_clip_files(path = ".", ...)
}
\arguments{
\item{path}{the path to search}

\item{...}{additional arguments passed to `fs::dir_tree`}

\item{dir_only}{logical indicating whether to return only directories or both files and directories}
}
\value{
a tree structure of files and directories
}
\description{
Copy the tree structure of files and directories to the clipboard
}
