% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_clip.R
\name{elm_clip_folders}
\alias{elm_clip_folders}
\title{Copy the tree structure of directories to the clipboard}
\usage{
elm_clip_folders(path = ".", ...)
}
\arguments{
\item{path}{the path to search}

\item{...}{additional arguments passed to `get_tree`}

\item{dir_only}{logical indicating whether to return only directories or both files and directories}
}
\value{
nothing, copies the file tree to the clipboard
}
\description{
Copy the tree structure of directories to the clipboard
}
