% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_sunburst.R
\name{elm_sunburst}
\alias{elm_sunburst}
\title{Generate a sunburst using the `sunburstR` package}
\usage{
elm_sunburst(data_tree_tibble, title = "Files or folders sunburst")
}
\arguments{
\item{data_tree_tibble}{A tibble of data tree information generated by `get_data_tree()`.}

\item{title}{A character string specifying the title of the sunburst.}
}
\value{
A sunburst of the files or directories in the specified directory.
}
\description{
This function generates a sunburst using the sunburstR package.
}
\examples{
\dontrun{
elm_sunburst(get_data_tree())
}
}
