% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_sunburst.R
\name{elm_sunburst_files}
\alias{elm_sunburst_files}
\title{Create a sunburst of files in a directory}
\usage{
elm_sunburst_files(path = ".", dir_only = F, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory of interest.}

\item{dir_only}{A logical value indicating whether to limit results to directories only. Default is FALSE.}

\item{...}{Additional arguments to be passed to `get_files_tibble()`.}
}
\value{
A sunburst of the files in the specified directory.
}
\description{
This function generates a sunburst of files in a given directory using the `elm_sunburst` function. By default, both files and directories are included, but files can be excluded by setting `dir_only` to TRUE.
}
\examples{
\dontrun{
elm_sunburst_files("~/Documents/")
}
}
