% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_sunburst.R
\name{elm_sunburst_folders}
\alias{elm_sunburst_folders}
\title{Create a sunburst of directories in a directory}
\usage{
elm_sunburst_folders(path = ".", dir_only = T, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory of interest.}

\item{dir_only}{A logical value indicating whether to limit results to directories only. Default is TRUE.}

\item{...}{Additional arguments to be passed to `get_files_tibble()`.}
}
\value{
A sunburst of the directories in the specified directory.
}
\description{
This function generates a sunburst of directories in a given directory using
the `elm_sunburst` function. By default, only directories are included, but
files can be included by setting `dir_only` to FALSE.
}
\examples{
\dontrun{
elm_sunburst_folders("~/Documents/")
}
}
