% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_treemap.R
\name{elm_treemap}
\alias{elm_treemap}
\title{Generate a treemap using the `highcharter` or `treemap` package}
\usage{
elm_treemap(
  files_tibble,
  title = "Files or folders treemap",
  interactive = TRUE
)
}
\arguments{
\item{files_tibble}{A tibble of file information generated by `get_files_tibble()`.}

\item{title}{A character string specifying the title of the treemap.}

\item{interactive}{A logical value indicating whether to display the treemap interactively using the `highcharter` package. Default is TRUE.}
}
\value{
A treemap of the files or directories in the specified directory.
}
\description{
This function generates a treemap using either the `highcharter` or `treemap` package, depending on the value of the `interactive` parameter.
}
\examples{
\dontrun{
elm_treemap(get_files_tibble("~/Documents/"))
}
}
