% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_treemap.R
\name{elm_treemap_files}
\alias{elm_treemap_files}
\title{Create a treemap of files in a directory}
\usage{
elm_treemap_files(path = ".", dir_only = F, interactive = TRUE, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory of interest.}

\item{dir_only}{A logical value indicating whether to limit results to directories only. Default is FALSE.}

\item{interactive}{A logical value indicating whether to display the treemap interactively using the `highcharter` package. Default is TRUE.}

\item{...}{Additional arguments to be passed to `get_files_tibble()`.}
}
\value{
A treemap of the files in the specified directory.
}
\description{
This function generates a treemap of files by size in a given directory using the
`elm_treemap` function. By default, both files and directories are included,
but files can be excluded by setting `dir_only` to TRUE.
}
\examples{
elm_treemap_files("~/Documents/")
}
