% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm_treemap.R
\name{elm_treemap_folders}
\alias{elm_treemap_folders}
\title{Create a treemap of directories in a directory}
\usage{
elm_treemap_folders(path = ".", dir_only = T, interactive = TRUE, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory of interest.}

\item{dir_only}{A logical value indicating whether to limit results to directories only. Default is TRUE.}

\item{interactive}{A logical value indicating whether to display the treemap interactively using the `highcharter` package. Default is TRUE.}

\item{...}{Additional arguments to be passed to `get_files_tibble()`.}
}
\value{
A treemap of the directories in the specified directory.
}
\description{
This function generates a treemap of directories in a given directory using the `elm_treemap` function. By default, only directories are included, but files can be included by setting `dir_only` to FALSE.
}
\examples{
elm_treemap_folders("~/Documents/")
}
