% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_data_tree}
\alias{get_data_tree}
\title{Get the data.tree object on files from a given path.}
\usage{
get_data_tree(path = ".", dir_only = FALSE, all = FALSE, ...)
}
\arguments{
\item{path}{A character string representing the path to the folder to be traversed.}

\item{dir_only}{A logical value indicating whether to only return directories in the tree.}

\item{all}{A logical value indicating whether to include hidden files/folders in the tree.}

\item{...}{Additional arguments to be passed to `get_files_tibble()`.}
}
\value{
A data.tree representing the file/folder tree.
}
\description{
Get the data.tree object on files from a given path.
}
\examples{
\dontrun{
get_tree()
}
}
