% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{get_error}
\alias{get_error}
\title{Compute model's errors with different number of neurons (nn)}
\usage{
get_error(object, ...)
}
\arguments{
\item{object}{An instance to the SLFN class.}

\item{n_sel}{The number of hidden neurons to be used.}

\item{h}{The transformed matrix H of dimensions [NxL].}

\item{y}{The output matrix of dimensions [Nxc].}

\item{h_val}{The transformed matrix H for the validation dataset
of dimensions [NxL].}

\item{y_val}{The output matrix for the validation dataset of dimensions [Nxc].}

\item{cv_rows}{The vector containing the selection of the data.}
}
\value{
The error of the ELM model
}
\description{
\code{computeError} calculate the error of a ELM model.
}

