% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files_tibble}
\alias{get_files_tibble}
\title{Get file information as a tibble}
\usage{
get_files_tibble(path = ".", dir_only = FALSE, all = FALSE, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory of interest.}

\item{dir_only}{A logical value indicating whether to limit results to directories only. Default is FALSE.}

\item{all}{A logical value indicating whether to include hidden files in the results. Default is FALSE.}

\item{...}{Additional arguments to be passed to `fs::dir_info()`.}
}
\value{
A tibble with file information for the specified path.
}
\description{
This function returns a tibble of file information for a given path. By default, it includes both files and directories, but can be limited to directories only by setting `dir_only` to TRUE.
}
\examples{
get_files_tibble("~/Documents/")
}
