% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_parent_folders}
\alias{get_parent_folders}
\title{Get parent folders from a path}
\usage{
get_parent_folders(path_split)
}
\arguments{
\item{path_split}{A character vector representing the split path of a file or directory.}
}
\value{
A character string representing the parent folder of the specified path.
}
\description{
This function returns the parent folder of a given path split into individual directory names.
}
\examples{
get_parent_folders(fs::path_split("~/Documents/folder/file.txt"))
}
