% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get the file/folder tree of a given path.}
\usage{
get_tree(path = ".", dir_only = FALSE, hidden = FALSE)
}
\arguments{
\item{path}{A character string representing the path to the folder to be traversed.}

\item{dir_only}{A logical value indicating whether to only return directories in the tree.}

\item{hidden}{A logical value indicating whether to include hidden files/folders in the tree.}
}
\value{
A character vector representing the file/folder tree.
}
\description{
Get the file/folder tree of a given path.
}
\examples{
\dontrun{
get_tree()
}
}
