% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_neurons.R, R/hiddenlayer-class.R, R/train_pruning.R
\docType{methods}
\name{add_neurons,hiddenlayer-method}
\alias{add_neurons,hiddenlayer-method}
\alias{hiddenlayer-class}
\alias{prune,hiddenlayer-method}
\alias{show,hiddenlayer-method}
\title{Class \code{hiddenlayer}}
\usage{
\S4method{add_neurons}{hiddenlayer}(object, ninputs, nn, act_fun, w_in, b)

\S4method{show}{hiddenlayer}(object)

\S4method{prune}{hiddenlayer}(object, n_sel)
}
\description{
A S4 class to encapsulate the parameters of the hidden layer in a SLFN
}
\section{Methods (by generic)}{
\itemize{
\item \code{add_neurons}: add neurons of the same type of activation function

\item \code{show}: show an object of class \code{hiddenlayer}

\item \code{prune}: prune/remove neurons from the hiddenlayer given an index of neurons
}}
\section{Slots}{

\describe{
\item{\code{act_fun}}{a \code{factor}, with the type of activation function used in each neurons. The levels
of the factor are the different type of activation function combined, while the length of
the factor equals to the number of neurons of the layer.}

\item{\code{w_in}}{a \code{matrix}, with the input weights. In the case of "rbf" neurons, it stands for the
centroids of each hidden neurons}

\item{\code{b}}{a \code{vector}, with the input bias. In the case of \emph{rbf} neurons, it stands
for the gamma values associated to each neuron.}
}}

