% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{loadSLFN,SLFN-method}
\alias{loadSLFN,SLFN-method}
\title{Load a SLFN
Compute the projection of the matrix H for a particular X}
\usage{
\S4method{loadSLFN}{SLFN}(object, filename = "", path = "")
}
\arguments{
\item{object}{SLFN object to serialize}

\item{filename}{name of the file where the SLFN object is read from}

\item{path}{a character string with the path name to a directory}
}
\description{
Load a SLFN
Compute the projection of the matrix H for a particular X
}

