% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\name{loadSLFN}
\alias{loadSLFN}
\title{Load a SLFN model.}
\usage{
loadSLFN(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class to be updated.}

\item{filename}{The name of the file where the SLFN object is read from.}

\item{path}{A character string with the path name to a directory.}
}
\description{
\code{loadSLFN} reads the settings of a SLFN from a file.
}

