% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{mse,SLFN-method}
\alias{mse,SLFN-method}
\title{Error calculation}
\usage{
\S4method{mse}{SLFN}(object, Y, Yp, H = NULL)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}

\item{Yp}{a matrix of dimensions [Nxc]; output matrix}
}
\value{
error the value of the model error
}
\description{
Error calculation
}

