% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{postprocess}
\alias{postprocess}
\title{Post-process classfication predictions}
\usage{
postprocess(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class.}

\item{typePred}{The type of post-processing implemented:
\itemize{
\item "raw" No post-processing: ELM predictions. Continuos predictions that do not follow the definition of probability.
\item "prob" Probability-like predictions. The softmax method is used for transformation.
The predictions followed a probability function, as they are between 0 and 1 and sum 1.
\item "label" Predicted labels from the probability-like continuous values.
}}

\item{ml_threshold}{Threshold value to assign multi-label targets.}
}
\value{
The output matrix/vector with post-processed predtictors
}
\description{
Function to post-process the continuous prediction of a SLFN in classfication
}
\details{
In classification, labels for multi-class and multi-label cases are computed as follows. For multi-class, the output
with higher continous value is selected. For multi-label, all outputs above \code{ml_threshold} are selected.
}

