% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,elm-method}
\alias{predict,elm-method}
\title{\code{\link{predict}} method for \code{\linkS4class{elm}} objects}
\usage{
\S4method{predict}{elm}(object, x, class_output = "prob",
  ml_threshold = 0.5)
}
\arguments{
\item{object}{the instance to elm class}

\item{X}{The input matrix of dimensions [Nxd].}
}
\value{
A output matrix of predictions Yp with dimensions [Nxc].
}
\description{
\code{predict} gives the estimation of input data X for the ELM model
}

