% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{project,SLFN-method}
\alias{project,SLFN-method}
\title{Compute the projection of the matrix H for a particular X}
\usage{
\S4method{project}{SLFN}(object, X = NULL)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{X}{a matrix of dimensions [Nxd]; input matrix}
}
\value{
H a matrix of dimensions [NxL]; matrix after transformation
}
\description{
Compute the projection of the matrix H for a particular X
}

