% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\name{project}
\alias{project}
\title{Compute the matrix H from X}
\usage{
project(object, X = NULL)
}
\arguments{
\item{object}{}

\item{X}{a matrix of dimensions [Nxd]; input matrix}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}
}
\value{
H a matrix of dimensions [NxL]; matrix after transformation
}
\description{
Compute the matrix H from X
}

