% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train_pruning.R
\name{prune}
\alias{prune}
\title{SLFN pruning}
\usage{
prune(object, ...)
}
\arguments{
\item{object}{An instance to the SLFN class.}

\item{n_sel}{a vector, with the indexes of neurons kept after pruning}
}
\value{
A SLFN object
}
\description{
Prune a SLFN, given the index of neurons selected
}

