% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{rankNeurons,SLFN-method}
\alias{rankNeurons,SLFN-method}
\title{Return ranking of hidden neurons: random or OP.
L1 regularization}
\usage{
\S4method{rankNeurons}{SLFN}(object, nn, H = NULL, Y = NULL)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{H}{a matrix of dimensions [NxL] after transformation}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}
}
\value{
rank an array with the order of the neurons
}
\description{
Return ranking of hidden neurons: random or OP.
L1 regularization
}

