% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ranking.R
\name{rankNeurons}
\alias{rankNeurons}
\title{Rank hidden neurons}
\usage{
rankNeurons(object, ...)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{H}{a matrix of dimensions [NxL] after transformation}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}
}
\value{
rank an array with the order of the neurons
}
\description{
Return ranking of hidden neurons: random or OP.
L1 regularization
}

