% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{safe_copy}
\alias{safe_copy}
\title{Copy tree to clipboard and print to console.}
\usage{
safe_copy(tree, prefix = "\\t\\t")
}
\arguments{
\item{tree}{A character vector representing the file/folder tree to be copied.}

\item{prefix}{A character string to be added as prefix to each line of the tree.}
}
\value{
The function does not return any value.
}
\description{
Copy tree to clipboard and print to console.
}
\examples{
safe_copy(write_tree(get_tree()))
}
