% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{saveSLFN,SLFN-method}
\alias{saveSLFN,SLFN-method}
\title{Save a SLFN
Compute the projection of the matrix H for a particular X}
\usage{
\S4method{saveSLFN}{SLFN}(object, filename, path = "")
}
\arguments{
\item{object}{SLFN object to serialize}

\item{filename}{name of the file where the SLFN object is read from}

\item{path}{a character string with the path name to a directory}
}
\description{
Save a SLFN
Compute the projection of the matrix H for a particular X
}

