% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveload.R
\name{saveSLFN}
\alias{saveSLFN}
\title{Save a SLFN model.}
\usage{
saveSLFN(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class to serialize.}

\item{filename}{The name of the file where the SLFN object is read from.}

\item{path}{A character string with the path name to a directory.}
}
\description{
\code{saveSLFN} create a .rda file taht stores the object SLFN created.
}

