% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFN-class.R
\docType{methods}
\name{show,SLFN-method}
\alias{show,SLFN-method}
\title{Display a SLFN object.}
\usage{
\S4method{show}{SLFN}(object)
}
\arguments{
\item{object}{The SLFN object to be displayed.}
}
\description{
Display a SLFN object.
}

