% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ELMclass.R
\docType{methods}
\name{show.ELM}
\alias{show.ELM}
\title{Display a ELM object}
\usage{
\S4method{show}{ELM}(object)
}
\arguments{
\item{object}{The ELM object to be displayed}
}
\description{
Display a ELM object
}

