% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{show.SFLN}
\alias{show.SFLN}
\title{Display a SLFN object}
\usage{
\S4method{show}{SLFN}(object)
}
\arguments{
\item{object}{The SFLN object to be displayed}
}
\description{
Display a SLFN object
}

