% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{class}
\name{SLFN-class}
\alias{SLFN-class}
\alias{addNeurons,SLFN-method}
\alias{project,SLFN-method}
\title{Class "SFLN" of Single-hidden Layer Feed-forward Network}
\usage{
\S4method{addNeurons}{SLFN}(object, type, number, W = NULL, B = NULL)

\S4method{project}{SLFN}(object, X)
}
\description{
Description
}
\details{
Details
}
\section{Methods (by generic)}{
\itemize{
\item \code{addNeurons}: Add Neurons to SLFN

\item \code{project}: Project from the features space to neurons space. Build H [N x L]
}}
\section{Slots}{

\describe{
\item{\code{inputs}}{number of inputs}

\item{\code{outputs}}{number of outputs}

\item{\code{neurons}}{list with the description of hidden neurons.}

\item{\code{beta}}{weight outuput vector}
}}

