% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{solve_system}
\alias{solve_system}
\title{Solve the linear system H %*% Wout = Y - [NxL] %*% [Lxc] x= [Nxc]
  Use orthogonal projection - correlation matrices
The function \code{solveSystem} solves the linear system under the equation
HH * Wout = HT - [LxL] %*% [Lxc] = [Lxc].}
\usage{
solve_system(object, ...)
}
\arguments{
\item{H}{a matrix of dimensions [NxL] after transformation}

\item{solve}{logical; needs to be true to return Wout value}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}
}
\value{
Wout a matrix of dimensions [Lxc] with the output weights
}
\description{
Solve the linear system H %*% Wout = Y - [NxL] %*% [Lxc] x= [Nxc]
  Use orthogonal projection - correlation matrices
The function \code{solveSystem} solves the linear system under the equation
HH * Wout = HT - [LxL] %*% [Lxc] = [Lxc].
}

