% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\name{solveSystem}
\alias{solveSystem}
\title{Solve the linear system H %*% Wout = Y - [NxL] %*% [Lxc] x= [Nxc]
  Use orthogonal projection - correlation matrices
Solve the linear system HH * Wout = HT - [LxL] %*% [Lxc] = [Lxc]
similar to .proj_cpu (akusok).}
\usage{
solveSystem(object, ...)
}
\arguments{
\item{H}{a matrix of dimensions [NxL] after transformation}

\item{getWout}{logical; needs to be true to return Wout value}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}
}
\value{
Wout a matrix of dimensions [Lxc] with the output weights
}
\description{
Solve the linear system H %*% Wout = Y - [NxL] %*% [Lxc] x= [Nxc]
  Use orthogonal projection - correlation matrices
Solve the linear system HH * Wout = HT - [LxL] %*% [Lxc] = [Lxc]
similar to .proj_cpu (akusok).
}

