% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{train,SLFN-method}
\alias{train,SLFN-method}
\title{Training method}
\usage{
\S4method{train}{SLFN}(object, X = NULL, Y = NULL,
  structureSelection = FALSE, ranking = "random", validation = "none",
  classification = "none", folds = "10", classType = "single", ...)
}
\arguments{
\item{object}{SLFN object to serialize}

\item{X}{}

\item{Y}{}

\item{structureSelection}{}

\item{ranking}{}

\item{validation}{}

\item{classification}{}

\item{folds}{}

\item{classType}{}

\item{...}{}
}
\description{
Training method
}

