% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\name{train}
\alias{train}
\title{Train a SLFN}
\usage{
train(object, ...)
}
\arguments{
\item{object}{SLFN object to serialize}

\item{...}{}

\item{X}{a data matrix of dimensions [Nxd] with input data}

\item{Y}{vector/matrix of outputs [Nx1c]}

\item{modelStrSel}{logical}

\item{ranking}{type of neurons ranking \code{random} or \code{lars}}

\item{validation}{t}

\item{classification}{}

\item{folds}{}

\item{classType}{}
}
\description{
\code{train} fits all the parameters that include a SLFN given a set of
 input data (X, Y) and a training scheme
}

