% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train_pruning.R
\name{train_pruning}
\alias{train_pruning}
\title{train a ELM with pruning}
\usage{
train_pruning(object, ...)
}
\arguments{
\item{object}{An instance to the ELM class.}

\item{h}{a \code{matrix} of dimensions [n x l], with the outputs of the hidden layer.}

\item{y}{a \code{matrix} of dimensions [n x c] or a \code{vector} if c = 1, with the output values.}

\item{h_val}{a \code{matrix} of dimensions [n x l], with the outputs of the hidden layer for the validation data.}

\item{y_val}{a \code{matrix} of dimensions [n x c] or a \code{vector} if c = 1, with the output values.}

\item{cv_rows}{The vector containing the selection of the data.}
}
\value{
A trained ELM object
}
\description{
train a ELM with pruning. The function implements an optimization algorithm for determine the optimium
number of neurons
}

