% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper.R
\name{trainELM}
\alias{trainELM}
\title{Create and train an ELM object}
\usage{
trainELM(X = NULL, Y = NULL, Xv = NULL, Yv = NULL, neurons,
  structureSelection = FALSE, validation = "none", ranking = "random",
  alpha, bigdata = FALSE)
}
\arguments{
\item{a}{numeric vector with the number of hidden neurons added. Each element of the vector corresponds to}

\item{a}{character vector indicating the different type of activation functions in the hidden neurons added.}
}
\description{
Create and train an ELM object
}

