% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pruning.R
\name{trainPruning}
\alias{trainPruning}
\title{train a SLFN with pruning}
\usage{
trainPruning(object, ...)
}
\arguments{
\item{object}{An instance to the SLFN class.}

\item{H}{}

\item{Y}{}

\item{Hv}{}

\item{Yv}{}

\item{index}{}
}
\value{
A trained SLFN object
}
\description{
train a SLFN with pruning. The function implements an optimization algorithm for determine the optimium
number of neurons
}

