% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{trainPrunning,SLFN-method}
\alias{trainPrunning,SLFN-method}
\title{Training}
\usage{
\S4method{trainPrunning}{SLFN}(object, X, Y, Xv = NULL, Yv = NULL)
}
\description{
Training
}

