% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{write_tree}
\alias{write_tree}
\title{Write a file/folder tree to character vector.}
\usage{
write_tree(tree)
}
\arguments{
\item{tree}{A character vector representing the file/folder tree to be written.}
}
\value{
A character vector representing the file/folder tree.
}
\description{
Write a file/folder tree to character vector.
}
\examples{
write_tree(get_tree())
}
