% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{computeError}
\alias{computeError}
\title{Compute model's errors with different number of neurons (nn)}
\usage{
computeError(object, ...)
}
\arguments{
\item{object}{An instance to the SLFN class.}

\item{nSelected}{The number of hidden neurons to be used.}

\item{H}{}

\item{Y}{}

\item{Hv}{}

\item{Yv}{}

\item{index}{}
}
\value{
The error of the ELM model
}
\description{
\code{computeError} calculate the error of a ELM model.
}

