% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_mat_vec}
\alias{dist_mat_vec}
\title{Distance matrix - vector}
\usage{
dist_mat_vec(x, ref, type = "euclidean")
}
\arguments{
\item{ref}{The reference vector, a sample/point from a \emph{d}-space.}

\item{type}{The particular method to compute the distance:
\itemize{
\item \emph{euclidean}
\item \emph{manhattan}
}}

\item{X}{A matrix with \emph{N} samples from a \emph{d}-space.}
}
\value{
A vector of length \emph{N} with all distances
}
\description{
The function \code{dist_mat_vec} compute the distance between each sample of
matrix (row-wise) and a reference vector.
}

