% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance.R
\name{distMatVect}
\alias{distMatVect}
\title{Distance matrix - vector}
\usage{
distMatVect(matrix, refVector, type = "euclidean")
}
\arguments{
\item{matrix}{a matrix with \emph{N} samples from \emph{d}-space}

\item{refVector}{a vector from \emph{d}-space}

\item{type}{method used to compute the distance between points
\itemize{
\item \emph{euclidean}
\item \emph{...}
}}
}
\value{
A vector of length \emph{N} with all distances
}
\description{
Compute the distance between each sample of matrix row-wise and a reference vector
}
\details{
Details
}

