% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance.R
\name{distMatVect}
\alias{distMatVect}
\title{Distance matrix - vector}
\usage{
distMatVect(X, ref, type = "euclidean")
}
\arguments{
\item{X}{a matrix with \emph{N} samples/points from \emph{d}-space}

\item{ref}{a vector, a sample/poing from \emph{d}-space}

\item{type}{method used to compute the distance
\itemize{
\item \emph{euclidean}
\item \emph{...}
}}
}
\value{
A vector of length \emph{N} with all distances
}
\description{
Compute the distance between each sample of matrix (row-wise)
and a reference vector. More details:
}

