% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ELMclass.R
\docType{class}
\name{ELM-class}
\alias{ELM-class}
\title{Class "ELM"
 --> ../man/ELM-class.Rd
     ~~~~~~~~~~~~~~~~~~~~~~~}
\description{
Class "ELM"
 --> ../man/ELM-class.Rd
     ~~~~~~~~~~~~~~~~~~~~~~~
}
\keyword{classes}

