% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elm-class.R
\docType{class}
\name{elm-class}
\alias{elm}
\alias{elm-class}
\alias{initialize,elm-method}
\alias{show,elm-method}
\title{Class \code{elm}}
\usage{
\S4method{initialize}{elm}(.Object, inputs = 0, outputs = 0)

\S4method{show}{elm}(object)
}
\description{
A S4 class to represent an Extreme Learning Machine (ELM) model
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: initalize an object of class \code{elm}

\item \code{show}: display an object of class \code{elm}
}}
\section{Slots}{

\describe{
\item{\code{type}}{The type of model implemented:
\itemize{
\item "reg": regression problem.
\item "class_mc": multi-class: the sample belongs to 1 class out of n.
\item "class_ml": multi-label: the sample can belong to m classes out of n (m<n).
\item "class_w":  weigted classification
}}

\item{\code{tune}}{Parameter to define the model structure selection method implemented to tune
the model hyper-parameters
#' \itemize{
\item "none": no model selection
\item "pruning": pruning of neurons of the hidden layer: P-ELM, if "ridge = 0 &
ranking = "random", OP-ELM, if "ridge = 0 & ranking = lars", TROP-ELM, if ("ridge != 0 &
ranking = lars)
}}

\item{\code{ranking}}{A character to select the type of ranking implemented when prunning option
is selected.
\itemize{
\item "random" - random ranking
\item "lars" - ranking based on lars - L1 penalty
}}

\item{\code{results}}{The error used to evaluate model performance.
mse c(mse_train, mse_val)}

\item{\code{ridge}}{The regularization parameter used to include the L2 penalty the#'}

\item{\code{validation}}{The validation procedure used for developing the model.
#' \itemize{
\item "none" - no validation process  <<<<<<ANDRES<<<<<<
\item "v" - validation. Xv and Yv are required
\item "cv" - cross validation. The number of folds is required
\item "loo" - leave one out based on the PRESS statistic
}}

\item{\code{folds}}{The number number of folds for the cross-validation procedure.}

\item{\code{class_weights}}{numeric vector of length = number_of_classes with the weigths for weighted type}

\item{\code{batch}}{The size of the bacth in an adaptative ELM.}

\item{\code{time_exec}}{The time of calculation for training the model.}

\item{\code{bigdata}}{An logical parameter to select the kind of acceleration used in
case of solving big data problems.}
}}

