\name{elm-package}
\alias{elm-package}
\alias{elm}
\docType{package}
\title{What the package does (short line)}
\description{
More about what it does (maybe more than one line) A concise (1-5 lines) description of the package
}
\details{
\tabular{ll}{
Package: \tab elm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-10-13\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Ruben Urraca-Valle, Andres Sanz-Garcia
Maintainer: Andres Sanz Garcia <andres.sanz-garcia@helsinki.fi>
}
\references{Package source is in the help file for package.skeleton}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
