% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elm-package.R
\docType{package}
\name{elm-package}
\alias{elm}
\alias{elm-package}
\title{elm: extreme learning machines for regression and classification}
\description{
A series of functions to create ELM models (i.e., develop and run models)
in R for regression and classification tasks, also predict the outputs and
represents results in plots. The package is intended to be used in both small
and big datasets.
}
\details{
\tabular{ll}{ Package: \tab elm\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2015-12-10\cr License: \tab GPL\cr LazyLoad: \tab yes\cr }
Add more information here!
}
\author{
Ruben Urraca and Andres Sanz-Garcia

Maintainer: Andres Sanz-Garcia <andres.sanz-garcia at helsinki.fi>
}
\references{
\url{github.com/mugiro/elm}
}
\seealso{
Rcpp package
}
\keyword{package}

