\name{elm-package}
\alias{elm-package}
\alias{elm}
\docType{package}
\title{
Exact linear models
}
\description{
Implementation of exact test of linear regressions.
}
\details{
\tabular{ll}{
Package: \tab elm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-11-17\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Karl Schlag, Olivier Gossner, Gareth Liu-Evans and Oliver Reiter
}
\references{
Schlag, K. and Gossner, O. (2013), "Finite-Sample Exact Tests for Linear Regressions with Bounded Dependent Variables", Journal of Econometrics (forthcoming). 
}
\seealso{
  \url{http://homepage.univie.ac.at/karl.schlag/research/statistics/}
}

