% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{mse}
\alias{mse}
\title{MSE error}
\usage{
mse(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class.}

\item{y}{The output matrix of dimensions [N x c] with number of columns
equivalent to the number of variables or classes.}

\item{yp}{The predicted output matrix of dimensions [N x c]; output matrix}

\item{x}{the input matrix [N x L]. Required for the LOO case}
}
\value{
The MSE error

If LOO is activated, the Allen's PRESS estimation is returned
}
\description{
Function to compute the mean squared error (MSE)
}
\details{
If LOO option is activated, the MSE Allen's PRESS formula is used and the data matrix X must be supplied.
}

