% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{mse}
\alias{mse}
\title{Errors calculation.}
\usage{
mse(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class.}

\item{Y}{The output matrix of dimensions [Nxc] with number of columns
equivalent to the number of variables or classes.}

\item{Yp}{The predicted output matrix of dimensions [Nxc]; output matrix}
}
\value{
error the value of the model error
}
\description{
\code{mse} determines the regression or classification error.
}

