% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{predict,SLFN-method}
\alias{predict,SLFN-method}
\title{\code{\link{predict}} method for \code{\linkS4class{SLFN}} objects}
\usage{
\S4method{predict}{SLFN}(object, X = NULL)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{X}{The input matrix of dimensions [Nxd].}
}
\value{
A output matrix of predictions Yp with dimensions [Nxc].
}
\description{
\code{predict} gives the estimation of input data X for the ELM model
}

