% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{predict,SLFN-method}
\alias{predict,SLFN-method}
\title{Predict targets for the given inputs X}
\usage{
\S4method{predict}{SLFN}(object, X = NULL)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{X}{a matrix of dimensions [Nxd]; input matrix}
}
\value{
Yp a matrix of dimensions [Nxc]; output matrix
}
\description{
Predict targets for the given inputs X
}

