% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{project}
\alias{project}
\title{Compute the projection of the matrix H for a particular X}
\usage{
project(object, ...)
}
\arguments{
\item{object}{the instance to SLFN class}

\item{X}{a matrix of dimensions [Nxd]; input matrix}
}
\value{
H a matrix of dimensions [NxL]; matrix after transformation
}
\description{
Compute the projection of the matrix H for a particular X
}

