% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{project}
\alias{project}
\title{Compute the projection of the matrix H for a particular X.}
\usage{
project(object, ...)
}
\arguments{
\item{object}{The instance to SLFN class.}

\item{X}{The input matrix of dimensions [Nxd].}

\item{typeDist}{The method to compute the distance. Default \code{euclidean}.}
}
\value{
A matrix H after transformation of dimensions [NxL].
}
\description{
\code{project} returns the projection of the matrix H.
}

