% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{class}
\name{SLFN-class}
\alias{SLFN-class}
\title{Class "SFLN" of Single-hidden Layer Feed-forward Network
 --> ../man/SLFN-class.Rd
     ~~~~~~~~~~~~~~~~~~~~~~~}
\description{
Class "SFLN" of Single-hidden Layer Feed-forward Network
 --> ../man/SLFN-class.Rd
     ~~~~~~~~~~~~~~~~~~~~~~~
}
\keyword{classes}

