% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{class}
\name{SLFN-class}
\alias{SLFN-class}
\alias{addNeurons,SLFN-method}
\alias{project,SLFN-method}
\title{Class "SFLN"}
\usage{
\S4method{addNeurons}{SLFN}(object, type, number, W = NULL, B = NULL)

\S4method{project}{SLFN}(object, X)
}
\description{
Class "SFLN" of Single-hidden Layer Feed-forward Network
}
\details{
Details
}
\section{Methods (by generic)}{
\itemize{
\item \code{addNeurons}: Add Neurons to SLFN

\item \code{project}: 
}}
\section{Slots}{

\describe{
\item{\code{inputs}}{number of inputs}

\item{\code{outputs}}{number of outputs}

\item{\code{neurons}}{a list with the description of hidden layer. The hidden layer can be composed by
neurons of different activation functions. Each element of the list groups neurons of the
same activation function. The element is named with the type of activation function and
contains the number of neurons (number), input weight vector (W) and biases (B) associated
to this group of neurons.}

\item{\code{Wout}}{a vector, the weight output vector, with the computed weights between the hidden
layer and the ouptut}

\item{\code{alpha}}{regularization parameter}

\item{\code{modelStrSel}}{character with the model structure selection method used
#' \itemize{
\item "none"
\item "prunning"
}}

\item{\code{ranking}}{character with the type of ranking implemented when prunning option is selected
\itemize{
\item "random" - random ranking
\item "lasso" - ranking based on LASSO - L1 penalty
}}

\item{\code{validation}}{validation procedure used for model structure
#' \itemize{
\item "V" - validation. Xv and Yv are required
\item "CV" - cross validation. The number of folds is required
\item "LOO" - leave one out based on the PRESS statistic
}}

\item{\code{folds}}{numeric, with the number of folds for cross-validation (CV)}

\item{\code{classification}}{numeric, with the type of classifation implemented:
\itemize{
\item "none" - regression
\item "sc" - single class: binary classification problem
\item "mc" multi-class: the sample belongs to 1 class out of n
\item "ml" multi-label: the sample can belong to m classes out of n (m<n)
\item "w"  weigted
}}
}}

