% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{solveSystem,SLFN-method}
\alias{solveSystem,SLFN-method}
\title{Solve the linear system H %*% beta = Y - [NxL] %*% [Lxc] = [Nxc]
  Use orthogonal projection - correlation matrices
Solve the linear system HH * beta = HT - [LxL] %*% [Lxc] = [Lxc]
similar to .proj_cpu (akusok).}
\usage{
\S4method{solveSystem}{SLFN}(object, H, Y, getBeta = TRUE)
}
\arguments{
\item{H}{a matrix of dimensions [NxL] after transformation}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}

\item{getBeta}{logical; needs to be true to return beta value}
}
\value{
beta a matrix of dimensions [Lxc] with the output weights
}
\description{
Solve the linear system H %*% beta = Y - [NxL] %*% [Lxc] = [Nxc]
  Use orthogonal projection - correlation matrices
Solve the linear system HH * beta = HT - [LxL] %*% [Lxc] = [Lxc]
similar to .proj_cpu (akusok).
}

