% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\name{solveSystem}
\alias{solveSystem}
\title{Solve the linear system H*beta = Y. Solve the system with otrhogonal
projection - correlation matrices HH*beta=HT similar to .proj_cpu (akusok).
May be it should be an S3 method...   params(object) <- newvalue}
\usage{
solveSystem(H, Y, getBeta = TRUE, ...)
}
\arguments{
\item{H}{a matrix of dimensions [NxL] after transformation}

\item{Y}{a matrix of dimensions [Nxc] - output matrix (columns = nº variables or classes)}

\item{getBeta}{logical; needs to be true to return beta value}
}
\value{
beta a matrix of dimensions [Lxc] with the output weights
}
\description{
Solve the linear system H*beta = Y. Solve the system with otrhogonal
projection - correlation matrices HH*beta=HT similar to .proj_cpu (akusok).
May be it should be an S3 method...   params(object) <- newvalue
}

