% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SLFNclass.R
\docType{methods}
\name{train,SLFN-method}
\alias{train,SLFN-method}
\title{Train a SLFN}
\usage{
\S4method{train}{SLFN}(object, X, Y, Xv, Yv, structureSelection = FALSE,
  ranking = "random", validation = "none", folds = "10",
  classification = "none", ...)
}
\arguments{
\item{object}{SLFN object to serialize}

\item{X}{a data matrix of dimensions [Nxd] with input data}

\item{Y}{vector/matrix of outputs [Nx1c]}

\item{structureSelection}{logical}

\item{ranking}{type of neurons ranking \code{random} or \code{LASSO}}

\item{validation}{}

\item{folds}{}

\item{classification}{}

\item{...}{}

\item{classType}{}
}
\description{
Training a SLFN given a set of input data (X, Y) and a training scheme
}

