% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a SLFN}
\usage{
train(object, ...)
}
\arguments{
\item{object}{SLFN object to serialize}

\item{...}{None to use until now}

\item{x}{a data matrix of dimensions [Nxd] with input data}

\item{y}{vector/matrix of outputs [Nx1c]}

\item{modelStrSel}{logical Select the pruning for reduce model's size.}

\item{ranking}{Type of neurons ranking \code{random} or \code{lars}.}

\item{validation}{Method to validate the model developed
\itemize{
\item "none" - no validation process
\item "V" - validation. Xv and Yv are required
\item "CV" - cross validation. The number of folds is required
\item "LOO" - leave one out based on the PRESS statistic
}}

\item{folds}{Number of folds defined for the cross-validation procedure}

\item{class_weights}{numeric vector of length = number_of_classes
with the weigths for weighted classification}

\item{classification}{"none"/"mc"/"ml"/"w"}
}
\description{
\code{train} fits all the parameters that include a SLFN given a set of
 input data (X, Y) and a training scheme
}

