% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{trainELM}
\alias{trainELM}
\title{trainELM}
\usage{
trainELM(X, Y, Xv = NULL, Yv = NULL, nType, nNumber, W = NULL, B = NULL,
  structureSelection = FALSE, validation = "none", ranking = "random",
  alpha, bigdata = FALSE, ...)
}
\arguments{
\item{X}{The input data values in a matrix or vector}

\item{Y}{The output data values in a matrix or vector}

\item{Xv}{The input data values in a matrix or vector for performing a simple
validation procedure}

\item{Yv}{The output data values in a matrix or vector for performing a
simple validation procedure}

\item{nType}{The types of activation functions used in a vector}

\item{W}{A list of suitable matrix with the input weight vectors or centroids
(rbf) per type of activation function}

\item{B}{A list of suitable vector with the input biases or sigmas (rbf)
per type of activation function}

\item{structureSelection}{A numeric vector with the number of hidden neurons added.}

\item{...}{Optional additional parameters. None are used at present.}

\item{nNeurons}{A vector containing the number of hidden neurons per type of
activation function}
}
\value{
An object of class \code{"SLFN"} with the model developed
}
\description{
\code{trainELM} creates and trains an object of class SLFN for given X and Y.
}
\details{
This function is a wrapper for summarizing several actions requiered when
creating and adjusting an ELM model. The particular steps are listed below.
\enumerate{
\item Creates the SLFN object by calling \code{new()}.
\item Adds the different hidden neurons by making sequential calls to \code{addNeurons()},
 one call per each type of activation function defined.
\item Trains the SLFN and obtaines the output weigth vector by calling \code{train()}.
}
}
\examples{
X = seq(0, 7, 0.01)
Y = sin(X)
a = trainELM(X = X, Y = Y, nType = "sigmoid", nNumber = 20)
a = trainELM(X = X, Y = Y, nType = c("sigmoid", "tanH"), nNumber = c(20, 10))
}

